pyproject.toml
[project]
name = "trip-assistant"
version = "0.1.0"
description = "AI-powered travel assistant"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "langchain-google-genai",
    "langgraph",
    "pydantic",
    "langchain-tavily"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"